githubauth: split app and installation into separate files
