+++ githubauth/app_test.go
@@ -15,7 +15,6 @@
 package githubauth
 
 import (
-	"context"
 	"crypto"
 	"crypto/rand"
 	"crypto/rsa"
@@ -23,9 +22,7 @@ import (
 	"crypto/x509"
 	"encoding/base64"
 	"encoding/pem"
-	"fmt"
 	"net/http"
-	"net/http/httptest"
 	"strings"
 	"testing"
 	"time"
@@ -234,302 +231,6 @@ func TestApp_OAuthAppTokenSource(t *testing.T) {
 	}
 
 	if got, want := oauthToken.AccessToken, token; got != want {
-		t.Errorf("expected %q to be %q", got, want)
-	}
-}
-
-func TestApp_AccessToken(t *testing.T) {
-	t.Parallel()
-
-	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	cases := []struct {
-		name     string
-		handler  http.HandlerFunc
-		req      *TokenRequest
-		expToken string
-		expErr   string
-	}{
-		{
-			name: "success",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(201)
-				fmt.Fprint(w, `{"token": "ghs_expectedtoken"}`)
-			},
-			req: &TokenRequest{
-				Repositories: []string{"my-repo"},
-				Permissions:  map[string]string{"issues": "write"},
-			},
-			expToken: "ghs_expectedtoken",
-		},
-		{
-			name: "success_empty_repositories",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(201)
-				fmt.Fprint(w, `{"token": "ghs_expectedtoken"}`)
-			},
-			req: &TokenRequest{
-				Repositories: []string{},
-				Permissions:  map[string]string{"issues": "write"},
-			},
-			expToken: "ghs_expectedtoken",
-		},
-		{
-			name: "fails_missing_repositories",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(201)
-			},
-			req: &TokenRequest{
-				Repositories: nil,
-				Permissions:  map[string]string{"issues": "write"},
-			},
-			expErr: "requested repositories cannot be nil",
-		},
-		{
-			name: "not_201",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(200)
-			},
-			req: &TokenRequest{
-				Repositories: []string{"my-repo"},
-				Permissions:  map[string]string{"issues": "write"},
-			},
-			expErr: "expected 200 to be 201",
-		},
-		{
-			name: "201_not_json",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(201)
-				fmt.Fprint(w, `this is not json`)
-			},
-			req: &TokenRequest{
-				Repositories: []string{"my-repo"},
-				Permissions:  map[string]string{"issues": "write"},
-			},
-			expErr: "failed to parse response as json",
-		},
-		{
-			name: "201_empty_body",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(201)
-			},
-			req: &TokenRequest{
-				Repositories: []string{"my-repo"},
-				Permissions:  map[string]string{"issues": "write"},
-			},
-			expErr: "failed to parse response as json",
-		},
-	}
-
-	for _, tc := range cases {
-		tc := tc
-
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-
-			ctx := context.Background()
-
-			mux := http.NewServeMux()
-			mux.Handle("/app/installations/123", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				fmt.Fprintf(w, `{"access_tokens_url": "http://%s/app/installations/123/access_tokens"}`, r.Host)
-			}))
-			mux.Handle("/app/installations/123/access_tokens", tc.handler)
-			srv := httptest.NewServer(mux)
-			t.Cleanup(srv.Close)
-
-			app, err := NewApp("my-app-id", privateKey, WithBaseURL(srv.URL))
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			installation, err := app.InstallationForID(ctx, "123")
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			token, err := installation.AccessToken(ctx, tc.req)
-			if diff := testutil.DiffErrString(err, tc.expErr); diff != "" {
-				t.Errorf(diff)
-			}
-
-			if got, want := token, tc.expToken; got != want {
-				t.Errorf("expected %q to be %q", got, want)
-			}
-		})
-	}
-}
-
-func TestApp_SelectedReposTokenSource(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-
-	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	mux := http.NewServeMux()
-	mux.Handle("/app/installations/123", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		fmt.Fprintf(w, `{"access_tokens_url": "http://%s/app/installations/123/access_tokens"}`, r.Host)
-	}))
-	mux.Handle("/app/installations/123/access_tokens", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(201)
-		fmt.Fprint(w, `{"token": "ghs_expectedtoken"}`)
-	}))
-	srv := httptest.NewServer(mux)
-	t.Cleanup(srv.Close)
-
-	app, err := NewApp("my-app-id", privateKey, WithBaseURL(srv.URL))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	installation, err := app.InstallationForID(ctx, "123")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	src := installation.SelectedReposTokenSource(map[string]string{"issues": "write"}, "my-repo")
-	githubToken, err := src.GitHubToken(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if got, want := githubToken, "ghs_expectedtoken"; got != want {
-		t.Errorf("expected %q to be %q", got, want)
-	}
-}
-
-func TestApp_AccessTokenAllRepos(t *testing.T) {
-	t.Parallel()
-
-	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	cases := []struct {
-		name     string
-		handler  http.HandlerFunc
-		req      *TokenRequestAllRepos
-		expToken string
-		expErr   string
-	}{
-		{
-			name: "success",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(201)
-				fmt.Fprint(w, `{"token": "ghs_expectedtoken"}`)
-			},
-			req: &TokenRequestAllRepos{
-				Permissions: map[string]string{"issues": "write"},
-			},
-			expToken: "ghs_expectedtoken",
-		},
-		{
-			name: "not_201",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(200)
-			},
-			expErr: "expected 200 to be 201",
-		},
-		{
-			name: "201_not_json",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(201)
-				fmt.Fprint(w, `this is not json`)
-			},
-			expErr: "failed to parse response as json",
-		},
-		{
-			name: "201_empty_body",
-			handler: func(w http.ResponseWriter, r *http.Request) {
-				w.WriteHeader(201)
-			},
-			expErr: "failed to parse response as json",
-		},
-	}
-
-	for _, tc := range cases {
-		tc := tc
-
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-
-			ctx := context.Background()
-
-			mux := http.NewServeMux()
-			mux.Handle("/app/installations/123", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				fmt.Fprintf(w, `{"access_tokens_url": "http://%s/app/installations/123/access_tokens"}`, r.Host)
-			}))
-			mux.Handle("/app/installations/123/access_tokens", tc.handler)
-			srv := httptest.NewServer(mux)
-			t.Cleanup(srv.Close)
-
-			app, err := NewApp("my-app-id", privateKey, WithBaseURL(srv.URL))
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			installation, err := app.InstallationForID(ctx, "123")
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			token, err := installation.AccessTokenAllRepos(ctx, tc.req)
-			if diff := testutil.DiffErrString(err, tc.expErr); diff != "" {
-				t.Errorf(diff)
-			}
-
-			if got, want := token, tc.expToken; got != want {
-				t.Errorf("expected %q to be %q", got, want)
-			}
-		})
-	}
-}
-
-func TestApp_AllReposTokenSource(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-
-	privateKey, err := rsa.GenerateKey(rand.Reader, 2048)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	mux := http.NewServeMux()
-	mux.Handle("/app/installations/123", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		fmt.Fprintf(w, `{"access_tokens_url": "http://%s/app/installations/123/access_tokens"}`, r.Host)
-	}))
-	mux.Handle("/app/installations/123/access_tokens", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(201)
-		fmt.Fprint(w, `{"token": "ghs_expectedtoken"}`)
-	}))
-	srv := httptest.NewServer(mux)
-	t.Cleanup(srv.Close)
-
-	app, err := NewApp("my-app-id", privateKey, WithBaseURL(srv.URL))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	installation, err := app.InstallationForID(ctx, "123")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	src := installation.AllReposTokenSource(map[string]string{"issues": "write"})
-	githubToken, err := src.GitHubToken(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if got, want := githubToken, "ghs_expectedtoken"; got != want {
 		t.Errorf("expected %q to be %q", got, want)
 	}
 }
